Changes to support multi-hypervisor.
Also to support testing on the OIL network, where I have to use an
overlay network.

This doesn't actually work yet - OpenStack can be setup with both
kvm and lxd hypervisors, but juju seems to ignore the constraints
(bug #1600404).

Also, when generating metadata, images get overwritten in the
simplestreams data so only the latest one added is there
(bug #1600054).